Make sure values to be compared are unsigned integers
It seems safer to make sure that input values are unsigned int before
casting them to unsigned long long and making the comparison.

+++ common.h
@@ -259,7 +259,7 @@ int get_wwnid_from_pretty
 int check_dir
 	(char *);
 void check_overflow
-	(unsigned long long, unsigned long long, unsigned long long);
+	(unsigned int, unsigned int, unsigned int);
 
 #ifndef SOURCE_SADC
 int count_bits
